feat: add support for secrets in CDK apps
Adds a construct that evaluates to a Secrets Manager secret value (not
SSM because CloudFormation does not yet support SSMPS SecureStrings).

In addition, a toolkit command has been added which can list all the secrets
used in the application and write values for them to Secrets Manager.

No toolkit integ test yet, since integ tests do an 'npm install' for
construct libraries and the published version of the libraries do not
have the required construct.

Fixes #656.